+++ jest.config.js
@@ -3,9 +3,7 @@ module.exports = {
     preset: 'ts-jest',
     testEnvironment: 'node',
     testMatch: ['**/__tests__/**/*.test.ts'],
-    globals: {
-        'ts-jest': {
-            tsconfig: 'tsconfig.json'
-        }
-    }
+    transform: {
+        '^.+\\.tsx?$': ['ts-jest', { tsconfig: 'tsconfig.json' }],
+    },
 };