+++ src/EventSub/index.ts
@@ -1,0 +1,2 @@
+export * from './idempotent';
+export { default } from './idempotent';